removed 2 minor warnings

+++ amiga-mui/datatypescache.c
@@ -80,7 +80,7 @@ Object *LoadPicture(char *filename, struct Screen *scr)
 	if (o)
 	{
 		struct FrameInfo fri = {0};
-		DoMethod(o,DTM_FRAMEBOX,NULL,&fri,&fri,sizeof(struct FrameInfo),0);
+		DoMethod(o,DTM_FRAMEBOX,NULL,(ULONG)&fri,(ULONG)&fri,sizeof(struct FrameInfo),0);
 	
 		if (fri.fri_Dimensions.Depth>0)
 		{
